+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - New option `--ual` to `odood addons install` and `odood addons update` comands
 - New option `--additional-addon` to `odood test` command
+- New option `--with-price` and `--without-price` for `odood addons list` command
+- New option `--color=price` for `odood addons list` command
 
 ## Release 0.0.8 (2023-05-05)